move game statistics methods to new GameStats class
